requirements.txt
# Core dependencies
pymysql==1.1.2
python-dotenv==1.0.1
